Implements victory condition and game reset
Implements the victory condition by tracking enemies killed and comparing it to the total number of enemies to spawn.

This also fixes a bug where the game wouldn't end properly.

Adds a `CheckVictory` method to `EnemySpawner` that is called when the number of enemies killed equals the total number of enemies to spawn, displaying a "You Win!" message and stopping enemy spawning.

Also adds a reference to the `EnemySpawner` component to the `Enemy` class, to check for victory upon enemy death.

Refactors and removes unused game reset logic from `EnemySpawner` and streamlines game initialization.